Updated Readme

+++ TheoryAssignment/C/Assignment-1/Prog5.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int a = 0, b = 0, c = 0;
+    float a = 0, b = 0, c = 0;
     printf("Enter the sides : ");
-    scanf("%d %d %d", &a, &b, &c);
+    scanf("%f %f %f", &a, &b, &c);
     
     if (a+b>c && a+c>b && b+c>a)
     {
